trying to catch with an if staement

+++ deploy_scripts/stop_cron.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
+set -e
 
 # stop all running cron jobs
-sudo pkill crond
+if [[ pgrep crond ]]; then
+  sudo pkill crond
+fi
 
 # stop cron
 sudo systemctl stop crond
